pyproject.toml
[project]
name = "SuikaiLauncher.Core"
version = "0.1.0"
description = "简单的跨平台 Minecraft 启动器核心库"
authors = [
    {name = "boximengling", email = "boximengling@outlook.com"},
]
dependencies = ["aiohttp>=3.11.14", "keyring>=25.6.0", "msal[broker]<2,>=1.20", "pyjwt>=2.10.1", "pillow>=11.1.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
